Switch to using signed integers for Type::bits and Type::width for backward compatibility.
Add -Wsign-compare warning flag as we should have this on everywhere if it is on anywhere.


Former-commit-id: a648600ea1fa69872c3819c15318aa90bd165346

+++ src/IR.h
@@ -438,7 +438,7 @@ struct Call : public ExprNode<Call> {
             << "Value index out of range in call to halide function\n";
         internal_assert(func.has_pure_definition() || func.has_extern_definition())
             << "Call to undefined halide function\n";
-        return make(func.output_types()[idx], func.name(), args, Halide, func, idx, Buffer(), Parameter());
+        return make(func.output_types()[(size_t)idx], func.name(), args, Halide, func, idx, Buffer(), Parameter());
     }
 
     /** Convenience constructor for loads from concrete images */
